+++ src/PlaywrightSharp.Tests/BaseTests/PlaywrightSharpBrowserLoaderFixture.cs
@@ -6,7 +6,7 @@ namespace PlaywrightSharp.Tests.BaseTests
     /// <summary>
     /// This class setup a single browser instance for tests.
     /// </summary>
-    public class PlaywrightSharpBrowserLoaderFixture : PlaywrightSharpLoaderFixture, IDisposable
+    public class PlaywrightSharpBrowserLoaderFixture : IDisposable
     {
         internal static IBrowserApp BrowserApp { get; private set; }
         internal static IBrowser Browser { get; private set; }
